requirements.txt
torch
transformers
datasets
trl
unsloth
python-docx
PyMuPDF
pyyaml
tenacity
requests